backend/requirements.txt
# FairLens Backend

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.1

# ML / XAI
scikit-learn==1.5.0
transformers>=4.40.0
huggingface-hub>=0.23.0
torch>=2.2.0
sentencepiece>=0.1.99
accelerate>=0.30.0
openai>=1.30.0
google-genai>=1.0.0
xgboost==2.1.1
lightgbm==4.5.0
shap==0.45.1
umap-learn==0.5.6
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
networkx==3.3

pyarrow==16.1.0

# Google Cloud
google-cloud-bigquery==3.25.0
google-cloud-storage==2.17.0
google-cloud-aiplatform==1.71.1

# LangGraph / LangChain
langgraph==0.1.14
langchain-core==0.2.38
langchain-google-vertexai==1.0.6

# HTTP
httpx==0.28.1
aiohttp==3.9.5

# Utils
python-multipart==0.0.9
aiofiles==23.2.1

# PDF reports
reportlab==4.2.2

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
